# back-end/src/swagger.json
{
	"components": {
...
		"schemas": {
			"IArticle": {
				"properties": {
					"_id": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"publishTime": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"cover": {
						"type": "string"
					},
					"abstract": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"collectNum": {
						"type": "number",
						"format": "double"
					},
					"commentNum": {
						"type": "number",
						"format": "double"
					},
					"watchNum": {
						"type": "number",
						"format": "double"
					},
					"likeNum": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"IRes_string_": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"msg": {
						"type": "string"
					},
					"data": {
						"type": "string"
					}
				},
				"required": [
					"code",
					"msg",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IRes_IArticle-Array_": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"msg": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/IArticle"
						},
						"type": "array"
					}
				},
				"required": [
					"code",
					"msg",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGolbalResponse_IArticle-Array_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/IRes_IArticle-Array_"
					},
					{
						"$ref": "#/components/schemas/IRes_string_"
					}
				]
			},
...
			"IComment": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"articleId": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"parentId": {
						"type": "string"
					},
					"likeUsers": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"dislikeUsers": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"collectUsers": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"children": {
						"items": {
							"$ref": "#/components/schemas/IComment"
						},
						"type": "array"
					}
				},
				"required": [
					"userId",
					"articleId",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IRes_IComment_": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"msg": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/IComment"
					}
				},
				"required": [
					"code",
					"msg",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGolbalResponse_IComment_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/IRes_IComment_"
					},
					{
						"$ref": "#/components/schemas/IRes_string_"
					}
				]
			},
			"IRes_IComment-Array_": {
				"properties": {
					"code": {
						"type": "number",
						"format": "double"
					},
					"msg": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/IComment"
						},
						"type": "array"
					}
				},
				"required": [
					"code",
					"msg",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IGolbalResponse_IComment-Array_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/IRes_IComment-Array_"
					},
					{
						"$ref": "#/components/schemas/IRes_string_"
					}
				]
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "tsoa-project",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/articles/list": {
			"get": {
				"operationId": "GetList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGolbalResponse_IArticle-Array_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "keyWord",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageIndex",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/comment/upload": {
			"post": {
				"operationId": "CreatComment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGolbalResponse_IComment_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IComment"
							}
						}
					}
				}
			}
		},
		"/comment/{articleId}": {
			"get": {
				"operationId": "Get",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IGolbalResponse_IComment-Array_"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "articleId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageIndex",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}